requirements.txt
pypdf==5.1.0
azure-search==1.0.0b2
azure-search-documents==11.5.2
python-dotenv==1.0.1
pandas
openai
langchain-text-splitters
ragas
